+++ src/routes/docs/plugins/examples/Appcss.svelte
@@ -1,12 +1,19 @@
-```css
-@import 'tailwindcss';
-// ...
-@plugin "flowbite-typography";
+<script lang="ts">
+  import { Clipboard } from "$lib";
+  import { CheckOutline, ClipboardCleanSolid } from "flowbite-svelte-icons";
 
-@source "../node_modules/@flowbite-svelte-plugins/texteditor/dist";
+  let value = $state("");
+  let success = $state(false);
 
-@layer components {
-  /* texteditor */
+  function onclick(ev: MouseEvent): void {
+    const target = ev.target as HTMLElement;
+    const codeBlock = target.ownerDocument.querySelector("#css-block");
+    if (codeBlock) {
+      value = codeBlock.textContent || "";
+    }
+  }
+
+  const cssCode = `/* texteditor */
   :root {
     --white: #fff;
     --black: #2e2b29;
@@ -99,10 +106,6 @@
   }
 
   /* floating menu */
-  .tippy-box {
-    max-width: 475px !important;
-  }
-
   .floating-menu {
     display: flex;
     background-color: var(--gray-3);
@@ -173,12 +176,12 @@
     display: none;
   }
 
-  /* texteditor Details */
+  /* Details */
   .tiptap .details {
     display: flex;
     gap: 0.25rem;
     margin: 1.5rem 0;
-    border: 1px solid #d4c8c760;
+    border: 1px solid var(--gray-3);
     border-radius: 0.5rem;
     padding: 0.5rem;
   }
@@ -204,7 +207,7 @@
   }
 
   .tiptap .details > button:hover {
-    background-color: #dedad8d8;
+    background-color: var(--gray-3);
   }
 
   .tiptap .details > button::before {
@@ -240,11 +243,11 @@
   }
 
   /* drag handle */
-  [id^='drag-handle-'] ::selection{
+  [id^='drag-handle-'] ::selection {
     background-color: #70cff850;
   }
 
-  .dark [id^='drag-handle-'] .ProseMirror-hideselection *::selection{
+  .dark [id^='drag-handle-'] .ProseMirror-hideselection *::selection {
     background-color: #e3508950 !important;
   }
 
@@ -309,15 +312,15 @@
     height: 1.25rem;
     content: '⠿';
     margin-top: 0.3rem;
-    top: 1rem !important;
+    /* top: 1rem !important; */
     font-weight: 700;
     cursor: grab;
     background: #0d0d0d10;
     color: #0d0d0d;
     border-radius: 0.25rem;
-    opacity: 1 !important;
+    /* opacity: 1 !important;
     visibility: visible !important;
-    pointer-events: auto !important;
+    pointer-events: auto !important; */
   }
 
   .dark [id^='drag-handle-'] .drag-handle {
@@ -350,8 +353,8 @@
   }
 
   /*  mention and emoji */
-  .tippy-box .tippy-content .dropdown-menu button,
-  .tippy-box .tippy-content .mention-dropdown button {
+  .emoji-suggestion-popup .dropdown-menu button,
+  .mention-suggestion-popup .mention-dropdown button {
     align-items: center;
     background-color: transparent;
     display: flex;
@@ -360,29 +363,31 @@
     width: 100%;
   }
 
-  .tippy-box .tippy-content .dropdown-menu button:hover,
-  .tippy-box .tippy-content .dropdown-menu button:hover.is-selected,
-  .tippy-box .tippy-content .mention-dropdown button:hover,
-  .tippy-box .tippy-content .mention-dropdown button:hover.is-selected {
+  .emoji-suggestion-popup .dropdown-menu button:hover,
+  .emoji-suggestion-popup .dropdown-menu button:hover.is-selected,
+  .mention-suggestion-popup .mention-dropdown button:hover,
+  .mention-dropdown button:hover.is-selected {
     background-color: rgba(61, 37, 20, 0.12);
   }
 
-  .tippy-box .tippy-content .dropdown-menu button.is-selected,
-  .tippy-box .tippy-content .mention-dropdown button.is-selected {
+  .emoji-suggestion-popup .dropdown-menu button.is-selected,
+  .mention-suggestion-popup .mention-dropdown button.is-selected {
     background-color: rgba(61, 37, 20, 0.08);
   }
 
-  .tippy-box .tippy-content .dropdown-menu button img {
+  .emoji-suggestion-popup .dropdown-menu button img,
+  .mention-suggestion-popup .dropdown-menu button img {
     height: 1em;
     width: 1em;
   }
 
-  .tippy-box {
+  .emoji-suggestion-popup,
+  .mention-suggestion-popup {
     transform-origin: center !important;
   }
 
-  .tippy-box .tippy-content .dropdown-menu,
-  .tippy-box .tippy-content .mention-dropdown {
+  .emoji-suggestion-popup .dropdown-menu,
+  .mention-suggestion-popup .mention-dropdown {
     background: #fff;
     border: 1px solid rgba(61, 37, 20, 0.05);
     border-radius: 0.7rem;
@@ -429,7 +434,7 @@
     pointer-events: none;
   }
 
-  .dark p.is-editor-empty:first-child::before{
+  .dark p.is-editor-empty:first-child::before {
     color: #666;
   }
 
@@ -536,31 +541,42 @@
     margin-bottom: 0.15em;
   }
 
-  /* math */
-  .tiptap .Tiptap-mathematics-editor {
-    background: #202020;
-    color: #fff;
-    font-family: monospace;
-    padding: 0.2rem 0.5rem;
-  }
-
-  .tiptap .Tiptap-mathematics-render {
+  /* Mathematics extension styles */
+  .tiptap .tiptap-mathematics-render {
     padding: 0 0.25rem;
   }
 
-  .tiptap .Tiptap-mathematics-render--editable {
+  .tiptap .tiptap-mathematics-render--editable {
     cursor: pointer;
     transition: background 0.2s;
   }
 
-  .tiptap .Tiptap-mathematics-render--editable:hover {
+  .tiptap .tiptap-mathematics-render--editable:hover {
     background: #eee;
   }
 
-  .tiptap .Tiptap-mathematics-editor,
-  .tiptap .Tiptap-mathematics-render {
+  .tiptap .tiptap-mathematics-render {
     border-radius: 0.25rem;
+  }
+
+  .tiptap .tiptap-mathematics-render[data-type='inline-math'] {
     display: inline-block;
+  }
+
+  .tiptap .tiptap-mathematics-render[data-type='block-math'] {
+    display: block;
+    margin: 1rem 0;
+    padding: 1rem;
+    text-align: center;
+  }
+
+  .tiptap .tiptap-mathematics-render.inline-math-error,
+  .tiptap .tiptap-mathematics-render.block-math-error {
+    background: var(--red-light);
+    color: var(--red);
+    border: 1px solid var(--red-dark);
+    padding: 0.5rem;
+    border-radius: 0.25rem;
   }
 
   /* Task list specific styles */
@@ -739,5 +755,21 @@
     outline: 3px solid var(--purple);
     transition: outline 0.15s;
   }
-}
-```
+
+  /* confit test */
+  .my-custom-class {
+    border: 2px solid rgb(150, 238, 206);
+  }`;
+</script>
+
+<div class="relative rounded-lg bg-gray-50 p-4 dark:bg-gray-700">
+  <pre><code id="css-block">{cssCode}</code></pre>
+
+  <Clipboard color={success ? "alternative" : "light"} bind:value bind:success size="sm" class="absolute end-2 top-2 h-8 px-2.5 font-medium focus:ring-0" {onclick}>
+    {#if success}
+      <CheckOutline class="h-3 w-3" /> Copied
+    {:else}
+      <ClipboardCleanSolid class="h-3 w-3" /> Copy code
+    {/if}
+  </Clipboard>
+</div>