+++ src/routes/docs/plugins/examples/Invisible.svelte
@@ -4,6 +4,14 @@
 
   let editorInstance = $state<Editor | null>(null);
 
+  function getEditorContent() {
+    return editorInstance?.getHTML() ?? "";
+  }
+
+  function setEditorContent(content: string) {
+    editorInstance?.commands.setContent(content);
+  }
+
   const content = `
       <h1>
         This is a heading.
@@ -17,6 +25,6 @@
     `;
 </script>
 
-<TextEditor bind:editor={editorInstance} {content}>
+<TextEditor bind:editor={editorInstance} {content} contentprops={{ id: "invisible-ex" }}>
   <InvisibleButtonGroup editor={editorInstance} />
 </TextEditor>