+++ src/routes/docs/plugins/examples/Tasklist.svelte
@@ -4,12 +4,15 @@
 
   let editorInstance = $state<Editor | null>(null);
 
-  const content = `<ul data-type="taskList">
+  const content = `Lorem ipsum dolor sit amet consectetur adipisicing elit. Quasi veniam nulla impedit, fugit similique nihil deserunt velit ea, laboriosam sequi!
+        <ul data-type="taskList">
           <li data-type="taskItem" data-checked="true">A list item</li>
           <li data-type="taskItem" data-checked="false">And another one</li>
-        </ul>`;
+        </ul>
+  Lorem ipsum dolor sit amet consectetur adipisicing elit. Quasi veniam nulla impedit, fugit similique nihil deserunt velit ea, laboriosam sequi!
+      `;
 </script>
 
-<TextEditor bind:editor={editorInstance} {content}>
+<TextEditor bind:editor={editorInstance} {content} contentprops={{ id: "task-ex" }}>
   <TaskListButtonGroup editor={editorInstance} />
 </TextEditor>