+++ vite.config.ts
@@ -4,7 +4,7 @@ import examples from "mdsvexamples/vite";
 import path from "path";
 import pkg from "./package.json" with { type: "json" };
 import devtoolsJson from "vite-plugin-devtools-json";
-
+import tiptapPackage from "./node_modules/@tiptap/core/package.json";
 import tailwindcss from "@tailwindcss/vite";
 import { defineConfig } from "vite";
 
@@ -35,6 +35,7 @@ export default defineConfig({
 
   define: {
     __NAME__: JSON.stringify(pkg.name),
-    __VERSION__: JSON.stringify(pkg.version)
+    __VERSION__: JSON.stringify(pkg.version),
+    __TIPTAP__: JSON.stringify(tiptapPackage.version)
   }
 });